requirements.txt
opencv-python
mediapipe
torch
torchvision
pillow
scikit-learn
numpy